requirements.txt
flask
gunicorn
mysql-connector-python
psycopg2-binary
bcrypt
python-dotenv
